Cross build to sbt 2.x
