+++ sbt-pgp/src/sbt-test/sbt-pgp/skip/test
@@ -1,2 +1,2 @@
 > publishSigned
--$ exists target/scala-2.13/test_2.13-1.0.jar.asc
+> check
